requirements.txt
psycopg2-binary
gspread
oauth2client
python-dotenv
rich
jinja2